requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
psycopg2-binary
python-dotenv
passlib
python-jose[cryptography]
pydantic
python-multipart
email-validator
